requirements.txt
fastapi
uvicorn
motor
pydantic
redis
faststream[rabbit]
python-dotenv
